pyproject.toml
[project]
name = "api-faturas"
version = "0.1.0"
description = "Add your description here"
authors = [{ name = "José Roberto Ferreira" }]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.3.3",
    "pyodbc>=5.2.0",
    "python-decouple>=3.8",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "schedule>=1.2.2",
    "sqlalchemy>=2.0.43",
]

[dependency-groups]
dev = ["sqlacodegen>=3.0.0"]

[tool.ruff]
line-length = 120
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'
